refactor(core): extract wasm options and lib cache from lib.rs
