+++ crates/tsz-core/src/api/wasm/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod compiler_options;
