+++ crates/tsz-core/src/project/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod lib_cache;
